[SPARK-46327][PS][CONNECT][TESTS] Reorganize `SeriesStringTests`
### What changes were proposed in this pull request?
Reorganize `SeriesStringTests`

### Why are the changes needed?
test code clean up

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44257 from zhengruifeng/ps_test_ser_str.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
